Widen the collapsed-read count to uint32
The ``count`` column of the reads blob holds how many identical
mappings collapse into one equivalence key.  As uint16 it topped out at
65,535, but a deep, non-deduplicated Yewdell library puts 87,488 reads
on a single key: under numpy 2 collection dies with an OverflowError,
and under numpy 1.26 it is worse — the value silently wraps (87,488 ->
21,952) behind a DeprecationWarning, so the hottest keys are undercounted
with no error at all.

The blob is a pickled DataFrame, so the dtype travels with the data and
databases collected before this change still load; the read path already
widens to int64 on load and never re-narrows.

The multimap clamp in ``Locus`` was justified by this wrap dropping a
slot's count below its baseline.  Keep the floor — the baseline is summed
from the spilled alignments independently of the collapsed count, so the
two disagreeing must still never make the Poisson response negative — but
state the reason that actually survives.

+++ price2/data_collector.py
@@ -581,13 +581,16 @@ _WORKER_LOCI: list[Locus] = []
 #: run land on the same worker repeatedly, so the index is parsed once.
 _WORKER_BAM: dict[str, pysam.AlignmentFile] = {}
 
+#: Layout of the collapsed-reads blob.  ``count`` is how many identical
+#: mappings collapse into one key; on deep, non-deduplicated libraries a
+#: single key can exceed 65,535, so it must be wider than ``uint16``.
 _READS_COLUMNS = (
     ("is_first_iv", bool),
     ("start", np.uint32),
     ("end", np.uint32),
     ("untemplated_addition", bool),
     ("unique", bool),
-    ("count", np.uint16),
+    ("count", np.uint32),
 )
